chore(ui): split sidebar components
